Adjust .gitignore for assets folder;
This is a proposal to ignore files in the `LovrApp/assets` folder so
they don't clutter `git status`.  Additionally, the presence of a
`.gitignore` file can serve the purpose of the existing "donotdelete"
file so that git tracks the folder.

I also removed the root level ignore for the `oculussig` files since the
new gitignore takes care of that.